AST: Split off Effects.h from Decl.h
Also remove the hash and equality operations on ProtocolRethrowsRequirementList;
this is no longer needed now that the AnyValue type eraser is gone.
